Refactors and fixes to get use x on y working.
Moved enums to seperate directory, and string fns for them in there too.

Fixed quite a few name and description issues and added logs to try to
catch these in the future.

Restructured the text creation in the HUD to handle cases that were not
being handled properly.

Fixed a bunch of odd little bugs.