Ugh. Lots of weird changes.
First the dialog system has introduced a lot of change. And it came
right when I was trying to use this as a template for a new game. Not a
great idea. Biggest changes are that Justin switches to using Bark text
in the HUD instead of in the scene, attached to the player. Also there's
a lot more integrating the bark text as a timed task, and using NPCs.

In my documentation I was trying to think about what is needed when
using the game as a template. I did not want to make it difficult to use
as a base for another game. I also wanted to understand how I felt about
the hunt for pixel perfection. Turns out I don't care: I just want the
game to be lovely.

At present the docs are a mess, and the dialog system is coming along.